+++ tests/functional/test_model_api.py
@@ -3,7 +3,7 @@ Model query and navigation API.
 """
 from __future__ import unicode_literals
 import pytest  # noqa
-from textx import metamodel_from_str, get_children_of_type, \
+from textx import metamodel_from_str, get_children, get_children_of_type, \
     get_parent_of_type, get_model
 
 
@@ -31,6 +31,38 @@ model_str = """
 """
 
 
+def test_get_children():
+
+    metamodel = metamodel_from_str(grammar)
+    model = metamodel.model_from_str(model_str)
+
+    seconds_thirds = get_children(
+        lambda x: x.__class__.__name__ in ['Second', 'Third'], model)
+
+    assert len(seconds_thirds) == 8  # 3 seconds and 5 thirds
+    assert seconds_thirds[0].x == [23, 45, 56]
+    assert seconds_thirds[1].x == 'one'
+    assert seconds_thirds[-1].x == 'third'
+
+    # Children first
+    seconds_thirds = get_children(
+        lambda x: x.__class__.__name__ in ['Second', 'Third'], model,
+        children_first=True)
+
+    assert len(seconds_thirds) == 8  # 3 seconds and 5 thirds
+    assert seconds_thirds[0].x == 'one'
+    assert seconds_thirds[1].x == [23, 45, 56]
+    assert seconds_thirds[-1].x == 'third'
+
+    # Do not traverse seconds
+    seconds_thirds = get_children(
+        lambda x: x.__class__.__name__ in ['Second', 'Third'], model,
+        should_follow=lambda x: x.__class__.__name__ != 'Second')
+    assert len(seconds_thirds) == 3  # Only 3 thirds at the top of the model
+    assert seconds_thirds[0].x == 'first'
+    assert seconds_thirds[-1].x == 'third'
+
+
 def test_get_children_of_type():
 
     metamodel = metamodel_from_str(grammar)
@@ -42,9 +74,16 @@ def test_get_children_of_type():
         == set([a.x for a in thirds])
 
     # Test search in the part of the model
-    thirds = get_children_of_type("Third", model.a[1])
+    thirds = get_children_of_type('Third', model.a[1])
     assert len(thirds) == 1
     assert 'two' == list(thirds)[0].x
+
+    # Do not traverse seconds
+    thirds = get_children_of_type(
+        'Third', model, should_follow=lambda x: x.__class__.__name__ != 'Second')
+    assert len(thirds) == 3  # Only 3 thirds at the top of the model
+    assert thirds[0].x == 'first'
+    assert thirds[-1].x == 'third'
 
 
 def test_get_parent_of_type():